# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management Api Documentation",
        "description": "<b>Api Documentation For</b>, Services responsible for getting the list of all events, get a single event details, creating, updating & deleting an event in database as well as attendee module i.e. getting all attendees for an event, attend an event, get details of an event's attendee & unAttend an Event.",
        "contact": {
            "name": "Aadhar gaur",
            "email": "aadhar41@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://events-management-app.test/api/"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of all events",
                "description": "Get list of all events",
                "operationId": "getEventList",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "relationship data i.e. attendees,user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aut est minima."
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Ipsam iste hic ut repellat totam est esse ad."
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "2024-02-11 21:18:46"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/events?include=user,attendees": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Store/Create event in DB",
                "description": "Store/Create event in DB",
                "operationId": "createEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test event Title"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2023-12-01 15:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2023-12-31 15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Test event Title"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2023-12-01 15:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2023-12-31 15:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Error: Not Found. The route events could not be found."
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "format": "object",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "191"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aut est minima."
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Ipsam iste hic ut repellat totam est esse ad."
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "2023-12-16 13:47:06"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "2024-02-11 21:18:46"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "id": "3",
                                                "name": "Aut est minima.",
                                                "description": "Ipsam iste hic ut repellat totam est esse ad.",
                                                "start_time": "2023-12-16 13:47:06",
                                                "end_time": "2024-02-11 21:18:46"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}